Split annet/generators into separate types
